requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==39.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-render-block==0.9.2
django-templated-email==3.0.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
exceptiongroup==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
packaging==23.0
phonenumberslite==8.13.6
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zope.interface==5.5.2
